+++ src/main/java/dev/willyelton/origins/common/item/ScalpelItem.java
@@ -36,6 +36,7 @@ public class ScalpelItem extends Item {
                 }
             }
         }
+
         return super.interactLivingEntity(itemStack, player, target, type);
     }
 }